[refactor] 출석, 영수증, 홈, 랜딩, 회비 페이지 tsx 변경
