+++ src/components/LogoutButton.tsx
@@ -1,8 +1,7 @@
-import React from 'react';
 import styled from 'styled-components';
 
-import logout from '../assets/images/ic_logout_gray.svg';
-import handleLogout from '../utils/handleLogout';
+import logout from '@/assets/images/ic_logout_gray.svg';
+import handleLogout from '@/utils/handleLogout';
 
 const ImgButton = styled.div`
   display: flex;
@@ -13,8 +12,11 @@ const ImgButton = styled.div`
   margin-right: 5%;
 `;
 
-const LogoutButton = () => {
-  const confirmLogout = handleLogout();
+const LogoutButton: React.FC = () => {
+  const confirmLogout = () => {
+    handleLogout();
+  };
+
   return (
     <ImgButton onClick={confirmLogout}>
       <img src={logout} alt="logout" />
