+++ src/pages/Home.tsx
@@ -1,12 +1,11 @@
-import React from 'react';
 import styled, { ThemeProvider } from 'styled-components';
-import theme from '../styles/theme';
-import HomeMain from '../components/home/HomeMain';
-import HomeFooter from '../components/home/HomeFooter';
-import LogoutButton from '../components/LogoutButton';
-import UserAPI from '../service/UserAPI';
-import logo from '../assets/images/ic_logo.svg';
-import useCustomBack from '../router/useCustomBack';
+import theme from '@/styles/theme';
+import HomeMain from '@/components/home/HomeMain';
+import HomeFooter from '@/components/home/HomeFooter';
+import LogoutButton from '@/components/LogoutButton';
+import UserAPI from '@/service/UserAPI';
+import logo from '@/assets/images/ic_logo.svg';
+import useCustomBack from '@/router/useCustomBack';
 
 const Container = styled.div`
   display: flex;
@@ -36,11 +35,8 @@ const Footer = styled.footer`
   width: 100%;
 `;
 
-const Home = () => {
+const Home: React.FC = () => {
   useCustomBack('/home');
-  // const accessToken = localStorage.getItem('accessToken');
-  // const refreshToken = localStorage.getItem('refreshToken');
-  // console.log('home token', accessToken, refreshToken);
   return (
     <ThemeProvider theme={theme}>
       <UserAPI />
