+++ src/pages/Receipt.tsx
@@ -1,8 +1,8 @@
 import styled from 'styled-components';
-import AttendHeader from '../components/Attendance/AttendHeader';
-import ReceiptMain from '../components/Receipt/ReceiptMain';
-import { DuesProvider } from '../service/DuesContext';
-import useCustomBack from '../router/useCustomBack';
+import AttendHeader from '@/components/Attendance/AttendHeader';
+import ReceiptMain from '@/components/Receipt/ReceiptMain';
+import { DuesProvider } from '@/service/DuesContext';
+import useCustomBack from '@/router/useCustomBack';
 
 const Container = styled.div`
   display: flex;
@@ -12,7 +12,7 @@ const Container = styled.div`
   margin-bottom: 50px;
 `;
 
-const Receipt = () => {
+const Receipt: React.FC = () => {
   useCustomBack('/dues');
 
   return (
